Add cloud to file_operations module
