+++ app/src/test/java/com/amaze/filemanager/filesystem/cloud/CloudStreamSourceTest.java
@@ -38,6 +38,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.annotation.Config;
 
+import com.amaze.filemanager.file_operations.filesystem.cloud.CloudStreamSource;
 import com.amaze.filemanager.shadows.ShadowMultiDex;
 
 import android.os.Environment;
