+++ file_operations/build.gradle
@@ -34,7 +34,28 @@ android {
     }
 }
 
+ext {
+    robolectricVersion = '4.3.1'
+    glideVersion = '4.11.0'
+    sshjVersion = '0.26.0'
+    jcifsVersion = '2.1.3'
+    fabSpeedDialVersion = '3.1.1'
+    roomVersion = '2.2.5'
+    bouncyCastleVersion = '1.65'
+    awaitilityVersion = "3.1.6"
+    androidXTestVersion = "1.2.0"
+    junitVersion = "4.13"
+    slf4jVersion = "1.7.25"
+    mockitoVersion = "3.4.4"
+    androidBillingVersion = "2.1.0"
+}
+
 dependencies {
+    //SFTP
+    implementation "com.hierynomus:sshj:$sshjVersion"
+
+    //smb
+    implementation "eu.agno3.jcifs:jcifs-ng:$jcifsVersion"
 
     //TODO some libs are not needed
     implementation fileTree(dir: "libs", include: ["*.jar"])
