+++ file_operations/src/main/java/com/amaze/filemanager/file_operations/filesystem/smbstreamer/StreamSource.java
@@ -18,25 +18,33 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.amaze.filemanager.filesystem.cloud;
+package com.amaze.filemanager.file_operations.filesystem.smbstreamer;
 
+/** Created by Arpit on 06-07-2015. */
 import java.io.IOException;
 import java.io.InputStream;
 
-import com.amaze.filemanager.filesystem.streams.RandomAccessStream;
+import com.amaze.filemanager.file_operations.filesystem.streams.RandomAccessStream;
 
-/** Created by Vishal on 30-04-2017. */
-public class CloudStreamSource extends RandomAccessStream {
+import android.webkit.MimeTypeMap;
+
+import jcifs.smb.SmbFile;
+
+public class StreamSource extends RandomAccessStream {
+
+  protected String mime;
   protected long fp;
   protected String name;
-  private InputStream inputStream;
+  protected SmbFile file;
+  InputStream input;
 
-  public CloudStreamSource(String fileName, long length, InputStream inputStream) {
-    super(length);
+  public StreamSource(SmbFile file, long l) {
+    super(l);
 
     fp = 0;
-    this.name = fileName;
-    this.inputStream = inputStream;
+    mime = MimeTypeMap.getFileExtensionFromUrl(file.getName());
+    name = file.getName();
+    this.file = file;
   }
 
   /**
@@ -58,7 +66,8 @@ public class CloudStreamSource extends RandomAccessStream {
    */
   public void open() throws IOException {
     try {
-      if (fp > 0) inputStream.skip(fp);
+      input = file.getInputStream(); // new SmbFileInputStream(file, bufferSize, 1);
+      if (fp > 0) input.skip(fp);
     } catch (Exception e) {
       throw new IOException(e);
     }
@@ -66,40 +75,47 @@ public class CloudStreamSource extends RandomAccessStream {
 
   @Override
   public int read() throws IOException {
-    int read = inputStream.read();
+    int read = input.read();
     if (read != -1) fp++;
     return read;
   }
 
-  @Override
   public int read(byte[] bytes, int start, int offs) throws IOException {
-    int read = inputStream.read(bytes, start, offs);
+    int read = input.read(bytes, start, offs);
     fp += read;
     return read;
   }
 
   @Override
+  public void moveTo(long position) throws IllegalArgumentException {
+    if (position < 0 || length() < position) {
+      throw new IllegalArgumentException("Position out of the bounds of the file!");
+    }
+
+    fp = position;
+  }
+
+  @Override
   public void close() {
-    if (inputStream != null) {
+    if (input != null) {
       try {
-        inputStream.close();
+        input.close();
       } catch (IOException e) {
         e.printStackTrace();
       }
     }
   }
 
+  public String getMimeType() {
+    return mime;
+  }
+
   public String getName() {
     return name;
   }
 
-  @Override
-  public void moveTo(long position) {
-    if (position < 0 || length() < position) {
-      throw new IllegalArgumentException("Position out of the bounds of the file!");
-    }
-
-    fp = position;
+  public SmbFile getFile() {
+    return file;
   }
 
   @Override
